refactor(codegen): extract get_field_type + get_expr_type_name into type_helpers.rs (#113 phase 7)
