+++ src/codegen/mod.rs
@@ -5,6 +5,7 @@ mod generics;
 mod intrinsics;
 mod lvalues;
 pub mod transpiler;
+mod type_helpers;
 mod types;
 
 pub use compiler::Compiler;
